+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/GetIncidentsForDetectionConfigurationOptions.cs
@@ -25,7 +25,7 @@ namespace Azure.AI.MetricsAdvisor
         {
             StartTime = startTime;
             EndTime = endTime;
-            DimensionsToFilter = new ChangeTrackingList<DimensionKey>();
+            DimensionKeys = new ChangeTrackingList<DimensionKey>();
         }
 
         /// <summary>
@@ -39,10 +39,12 @@ namespace Azure.AI.MetricsAdvisor
         public DateTimeOffset EndTime { get; }
 
         /// <summary>
-        /// Filters the result by series. Only incidents detected in the time series groups specified will
-        /// be returned.
+        /// Filters the result by time series. Each element in this list represents a set of time series, and only
+        /// incidents detected in at least one of these sets will be returned. For each element, if all possible
+        /// dimensions are set, the key uniquely identifies a single time series for the corresponding metric. If
+        /// only a subset of dimensions are set, the key uniquely identifies a group of time series.
         /// </summary>
-        public IList<DimensionKey> DimensionsToFilter { get; }
+        public IList<DimensionKey> DimensionKeys { get; }
 
         /// <summary>
         /// If set, specifies the maximum limit of items returned in each page of results. Note:
@@ -55,7 +57,7 @@ namespace Azure.AI.MetricsAdvisor
         {
             DetectionIncidentFilterCondition filterCondition = new DetectionIncidentFilterCondition();
 
-            foreach (DimensionKey dimensionKey in DimensionsToFilter)
+            foreach (DimensionKey dimensionKey in DimensionKeys)
             {
                 filterCondition.DimensionFilter.Add(dimensionKey.Clone());
             }