+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/AnomalyDetection/MetricSingleSeriesDetectionCondition.cs
@@ -17,7 +17,7 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// <summary>
         /// Initializes a new instance of the <see cref="MetricSingleSeriesDetectionCondition"/> class.
         /// </summary>
-        /// <param name="seriesKey">The key that uniquely identifies the time series to which these conditions apply within a metric. Every dimension contained in the associated <see cref="DataFeed"/> must be assigned a value.</param>
+        /// <param name="seriesKey">The key that uniquely identifies the time series to which these conditions apply within a metric. All possible dimensions must be assigned a value.</param>
         /// <exception cref="ArgumentNullException"><paramref name="seriesKey"/> is <c>null</c>.</exception>
         public MetricSingleSeriesDetectionCondition(DimensionKey seriesKey)
         {
@@ -34,7 +34,7 @@ namespace Azure.AI.MetricsAdvisor.Models
 
         /// <summary>
         /// The key that uniquely identifies the time series to which these conditions apply within a metric.
-        /// Every dimension contained in the associated <see cref="DataFeed"/> must be assigned a value.
+        /// All possible dimensions must be assigned a value.
         /// </summary>
         public DimensionKey SeriesKey { get; set; }