+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/MetricFeedback/MetricPeriodFeedback.cs
@@ -20,8 +20,8 @@ namespace Azure.AI.MetricsAdvisor
         /// <param name="metricId">The identifier of the metric to which the <see cref="MetricPeriodFeedback"/> applies.</param>
         /// <param name="dimensionKey">
         /// A key that identifies a set of time series to which the <see cref="MetricPeriodFeedback"/> applies.
-        /// If all possible dimension columns are set, this key uniquely identifies a single time series
-        /// for the specified <paramref name="metricId"/>. If only a subset of dimension columns are set, this
+        /// If all possible dimensions are set, this key uniquely identifies a single time series
+        /// for the specified <paramref name="metricId"/>. If only a subset of dimensions are set, this
         /// key uniquely identifies a group of time series.
         /// </param>
         /// <param name="periodType">The <see cref="MetricPeriodType"/>.</param>
@@ -37,11 +37,11 @@ namespace Azure.AI.MetricsAdvisor
 
         /// <summary> Initializes a new instance of MetricPeriodFeedback. </summary>
         /// <param name="metricId"> metric unique id. </param>
-        /// <param name="dimensionFilter"> . </param>
+        /// <param name="feedbackFilter"> . </param>
         /// <param name="valueInternal"> . </param>
-        /// <exception cref="ArgumentNullException"> <paramref name="metricId"/>, <paramref name="dimensionFilter"/>, or <paramref name="valueInternal"/> is null. </exception>
-        internal MetricPeriodFeedback(string metricId, FeedbackFilter dimensionFilter, PeriodFeedbackValue valueInternal)
-            : base(metricId, dimensionFilter.DimensionKey)
+        /// <exception cref="ArgumentNullException"> <paramref name="metricId"/>, <paramref name="feedbackFilter"/>, or <paramref name="valueInternal"/> is null. </exception>
+        internal MetricPeriodFeedback(string metricId, FeedbackFilter feedbackFilter, PeriodFeedbackValue valueInternal)
+            : base(metricId, feedbackFilter.DimensionKey)
         {
             if (valueInternal == null)
             {