+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/MetricsAdvisorClient/AnomalyDetectionTests.cs
@@ -197,10 +197,10 @@ namespace Azure.AI.MetricsAdvisor.Tests
         {
             MetricsAdvisorClient client = GetMetricsAdvisorClient();
 
-            var columns = new Dictionary<string, string>() { { "name", "value" } };
+            var dimensions = new Dictionary<string, string>() { { "name", "value" } };
 
             var emptyList = new List<DimensionKey>();
-            var seriesKeys = new List<DimensionKey>() { new DimensionKey(columns) };
+            var seriesKeys = new List<DimensionKey>() { new DimensionKey(dimensions) };
 
             Assert.That(() => client.GetMetricEnrichedSeriesDataAsync(FakeGuid, null, default, default), Throws.InstanceOf<ArgumentNullException>());
             Assert.That(() => client.GetMetricEnrichedSeriesDataAsync(FakeGuid, emptyList, default, default), Throws.InstanceOf<ArgumentException>());
@@ -220,10 +220,10 @@ namespace Azure.AI.MetricsAdvisor.Tests
         {
             MetricsAdvisorClient client = GetMetricsAdvisorClient();
 
-            var columns = new Dictionary<string, string>() { { "name", "value" } };
+            var dimensions = new Dictionary<string, string>() { { "name", "value" } };
 
             var emptyList = new List<DimensionKey>();
-            var seriesKeys = new List<DimensionKey>() { new DimensionKey(columns) };
+            var seriesKeys = new List<DimensionKey>() { new DimensionKey(dimensions) };
 
             using var cancellationSource = new CancellationTokenSource();
             cancellationSource.Cancel();